Delete updateHand method and fix bug on checkWinner method
Remove updateHand() from the player hand given that it was unnecessary, there's no need to keep track of the cards on player's hand by saving them into an object's property. Also fix a bug on the checkWinner method so that now actually checks whether one player has won and stops the game if so.